Make ouroboros-consensus warning free
Fixes #1659.

We now use the following warning flags for all components in consensus:

    -Wall
    -Wcompat
    -Wincomplete-uni-patterns
    -Wincomplete-record-updates
    -Wpartial-fields
    -Widentities
    -Wredundant-constraints
    -Wmissing-export-lists

This commit fixes all warnings produced by these flags.

+++ ouroboros-consensus/ouroboros-consensus-test-infra/src/Test/Util/FS/Sim/MockFS.hs
@@ -628,7 +628,7 @@ hPutSome h toWrite =
     snip :: Int -> Int -> ByteString -> (ByteString, ByteString)
     snip n m bs = (a, c)
       where
-        (a, bc) = BS.splitAt (fromIntegral n) bs
+        (a, bc) = BS.splitAt n bs
         c       = BS.drop m bc
 
 -- | Truncate a file
